% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{boolean_matrix_to_integer_list}
\alias{boolean_matrix_to_integer_list}
\title{Map boolean matrix defined by \code{var}, as described in
\code{matrix_from}, to a list of vector of integers.}
\usage{
boolean_matrix_to_integer_list(df, var, name = NULL)
}
\arguments{
\item{df}{masked data}

\item{var}{symbolic variable used to represent the matrix}

\item{name}{column name of matrix strings}
}
\description{
Map boolean matrix defined by \code{var}, as described in
\code{matrix_from}, to a list of vector of integers.
}
