% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{encode_matrix}
\alias{encode_matrix}
\title{Encodes a matrix as a data frame with specified columns.}
\usage{
encode_matrix(mat, var)
}
\arguments{
\item{mat}{matrix}

\item{var}{the symbolic name of the matrix (prefix of column names)}
}
\value{
a tibble (data frame) encoding of a matrix
}
\description{
Encodes a matrix as a data frame with specified columns.
}
