% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{infer_num_comp}
\alias{infer_num_comp}
\title{Infer the number of components in a system from a data frame.}
\usage{
infer_num_comp(df, pat = NULL)
}
\arguments{
\item{df}{data frame object}
}
\value{
inferred number of components in \code{df}
}
\description{
A component in a system is something that, for the sake of the analysis,
cannot be decomposed into more parts.
}
