% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{is_md}
\alias{is_md}
\title{Test whether an object is a masked data (\code{tbl_md}).}
\usage{
is_md(x)
}
\arguments{
\item{x}{object to determine if masked data}
}
\description{
Test whether an object is a masked data (\code{tbl_md}).
}
