% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{matrix_from}
\alias{matrix_from}
\title{Obtains a matrix from specified columns in a data frame.}
\usage{
matrix_from(df, var)
}
\arguments{
\item{df}{data frame that contains the matrix}

\item{var}{the symbolic name of the matrix}
}
\value{
a matrix
}
\description{
An \code{nrow(df)}-by-\code{p} boolean matrix \code{var} is encoded in
data fame \code{df} as the columns \code{var.1},...,\code{var.p} or
\code{var1},...,\code{varp}.
}
\details{
A matrix will be returned with the appropriate ordering denoted by the
index, e.g., \code{a.2} will come before \code{a.4}. If a column labeled
\code{a.3} is missing, we just make \code{a.2} and \code{a.4} adjacent.
}
