% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md}
\alias{md}
\title{md}
\usage{
md(x)
}
\arguments{
\item{x}{An object to convert to masked data.}
}
\value{
A masked data tibble with additional attributes.
}
\description{
Constructor for masked data.
}
\details{
Converts an object `x` to masked data, which is a tibble (data frame) with
some extra attributes, such as the `latent` attribute to specify which
variables are latent in the model.
}
