% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_boolean_matrix_to_charsets}
\alias{md_boolean_matrix_to_charsets}
\title{md_boolean_matrix_to_charsets}
\usage{
md_boolean_matrix_to_charsets(df, setvar = "x", cname = NULL, drop_set = FALSE)
}
\arguments{
\item{df}{A data frame containing the Boolean matrix to be converted.}

\item{setvar}{A character string specifying the variable name prefix for the columns of the Boolean matrix in the input data frame.}

\item{cname}{A character string specifying the name of the new column to be added to the output data frame. If NULL, the value of 'setvar' will be used as the column name. Default is NULL.}

\item{drop_set}{A logical value indicating whether to drop the original columns of the Boolean matrix from the output data frame. Default is FALSE.}
}
\value{
A data frame with the string representation of the Boolean matrix added as a new column.
}
\description{
Convert a Boolean Matrix to charsets (character set representation).
}
\details{
This function takes a Boolean matrix and converts it to a string representation.
Each row of the matrix is represented as a set enclosed in curly braces.
The resulting string representation is added as a new column to the input data frame.
}
\examples{
# Create a sample data frame with a Boolean matrix
df <- data.frame(x1 = c(TRUE, FALSE, TRUE), x2 = c(FALSE, TRUE, FALSE))
# Convert the Boolean matrix to string representation
result <- md_boolean_matrix_to_charsets(df, setvar = "x", cname = "X")
print(result)
}
