% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_cand_contains}
\alias{md_cand_contains}
\title{Decorates masked data with whether candidate set contains failed component.}
\usage{
md_cand_contains(md)
}
\arguments{
\item{md}{masked data frame}
}
\description{
Takes masked data frame \code{md} with candidate set encoded as \code{x}
and a column \code{k} denoting component cause of failure and returns a
decorated masked data frame with a column \code{contains} that
denotes whether the candidate set contains the component cause of failure.
}
