% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_cand_sizes}
\alias{md_cand_sizes}
\title{Decorates masked data with candidate sizes}
\usage{
md_cand_sizes(md)
}
\arguments{
\item{md}{masked data frame}
}
\description{
Takes masked data frame \code{md} with candidate set encoded as \code{x}
and returns a decorated masked data frame with a column \code{w} that
denotes the size of the candidate sets. No new information is added, it
just counts the number of times that a row element of \code{x} is \code{TRUE}.
}
