% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_charsets_to_boolean_matrix}
\alias{md_charsets_to_boolean_matrix}
\title{md_charsets_to_boolean_matrix}
\usage{
md_charsets_to_boolean_matrix(df, cname, drop_set = FALSE, setvar = NULL)
}
\arguments{
\item{df}{data frame}

\item{cname}{character, name of column containing charsets}

\item{drop_set}{logical, if TRUE drop column cname}

\item{setvar}{character, prefix of matrix encoding}
}
\value{
Boolean matrix representing charsets
}
\description{
Function to convert the character vector column to a Boolean matrix
}
