% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_convert_boolean_matrix_to_charsets}
\alias{md_convert_boolean_matrix_to_charsets}
\title{md_convert_boolean_matrix_to_charsets}
\usage{
md_convert_boolean_matrix_to_charsets(mat)
}
\arguments{
\item{mat}{Boolean matrix representing a set whose elements are indexed by
the natural numbers}
}
\value{
a list of character vectors representing the sets
}
\description{
Function to convert a Boolean matrix to a vector of string sets (charsets)
It is the inverse operation of `md_convert_charsets_to_boolean_matrix`.
}
