% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_convert_boolean_matrix_to_list}
\alias{md_convert_boolean_matrix_to_list}
\title{md_convert_boolean_matrix_to_list}
\usage{
md_convert_boolean_matrix_to_list(mat)
}
\arguments{
\item{mat}{Boolean matrix}
}
\value{
list of integer vectors
}
\description{
Map Boolean matrix `mat` (each row represnts a set as a Boolean vector, where
TRUE indicates membership and FALSE otherwise) to a list of integer vectors.
}
