% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_convert_charsets_to_boolean_matrix}
\alias{md_convert_charsets_to_boolean_matrix}
\title{md_charsets_to_boolean_matrix}
\usage{
md_convert_charsets_to_boolean_matrix(xs)
}
\arguments{
\item{xs}{The list of integer sets}
}
\value{
Boolean matrix representing each of the integer sets
}
\description{
Function to convert the character vector column to a Boolean matrix
}
\examples{
xs <- c("{1,2}", "{2,3}", "{2}")
print(md_convert_charsets_to_boolean_matrix(xs))
##       [,1]  [,2]  [,3]
## [1,]  TRUE  TRUE  FALSE
## [2,]  FALSE TRUE  TRUE
## [3,]  FALSE TRUE  FALSE
}
