% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_convert_list_to_boolean_matrix}
\alias{md_convert_list_to_boolean_matrix}
\title{md_convert_list_to_boolean_matrix}
\usage{
md_convert_list_to_boolean_matrix(xs)
}
\arguments{
\item{xs}{List of integer vectors.}
}
\value{
Boolean matrix representation of `xs`
}
\description{
Map list of integer vectors (each list represents a set) to Boolean matrix,
where each row represents the corresponding set as a Boolean vector.
}
