% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_decode_boolean_matrix_as_list}
\alias{md_decode_boolean_matrix_as_list}
\title{Map Boolean matrix encoded by \code{var}, as described in
\code{md_decode_matrix} and \code{md_encode_matrix}, to a list of vector of
integers.}
\usage{
md_decode_boolean_matrix_as_list(df, var)
}
\arguments{
\item{df}{data frame}

\item{var}{symbolic variable used to represent Boolean matrix in \code{df}}
}
\description{
Map Boolean matrix encoded by \code{var}, as described in
\code{md_decode_matrix} and \code{md_encode_matrix}, to a list of vector of
integers.
}
