% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_decode_matrix}
\alias{md_decode_matrix}
\title{Decodes a matrix from specified columns in a data frame.}
\usage{
md_decode_matrix(df, var)
}
\arguments{
\item{df}{data frame that contains the matrix}

\item{var}{the symbolic name of the matrix}
}
\value{
a matrix
}
\description{
An \code{nrow(df)}-by-\code{p} matrix \code{var} is encoded in
data frame \code{df} with the columns \code{var.1},...,\code{var.p} or
\code{var1},...,\code{varp}.
}
\details{
A matrix will be returned with the appropriate ordering denoted by the
index, e.g., \code{a.2} will come before \code{a.4}. There should be no
gaps in the matrix indexes, e.g., if there is \code{a.4} then there must
also be \code{a.1,a.2,a.3}.
}
