% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_list_to_boolean_matrix}
\alias{md_list_to_boolean_matrix}
\title{Map list of integer vectors to Boolean matrix.}
\usage{
md_list_to_boolean_matrix(xs)
}
\arguments{
\item{xs}{List of integer vectors.}
}
\description{
Map list of integer vectors to Boolean matrix.
}
