% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_read_csv_with_meta}
\alias{md_read_csv_with_meta}
\title{Read a data frame table from a connection (e.g., url or filename).}
\usage{
md_read_csv_with_meta(
  file,
  read_meta = T,
  comment = "#",
  max_meta_lns = 1000,
  skip = 0
)
}
\arguments{
\item{file}{a path to a file, a connection, or literal data}

\item{read_meta}{whether to read in metadata to populate attributes}

\item{comment}{comment indicator, defaults to "#'}

\item{max_meta_lns}{limit metadata search to the indicated number of lines}

\item{skip}{a number indicating how many of the first lines to skip}
}
\description{
Read a data frame table from a connection (e.g., url or filename).
}
