% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_ttf.R
\name{md_series_ttf}
\alias{md_series_ttf}
\title{Generates time-to-failure (ttf) and component cause of failure for a series
system with the given data frame of component times-to-failure encoded by
the matrix columns prefixed with \code{var} in masked data frame \code{md}.}
\usage{
md_series_ttf(md, tau = NULL, var = "t")
}
\arguments{
\item{md}{a data frame with the indicated component times-to-failure}

\item{tau}{suspension time}

\item{var}{component ttf symbolic variable, defaults to \code{t}}
}
\description{
Generates time-to-failure (ttf) and component cause of failure for a series
system with the given data frame of component times-to-failure encoded by
the matrix columns prefixed with \code{var} in masked data frame \code{md}.
}
