% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_set_contains}
\alias{md_set_contains}
\title{md_set_contains}
\usage{
md_set_contains(df, setvar, elvar, cname = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{setvar}{A character string representing the column prefix for matrix encoding.}

\item{elvar}{A character string representing the column name for elements to test membership for in setvar.}

\item{cname}{An optional character string for the name of the new column denoting outcomes of membership tests. Defaults to contains(setvar,elvar).}
}
\value{
A decorated data frame with a new column cname indicating the membership test outcomes.
}
\description{
Decorates a data frame by adding a new column indicating whether a set
(represented as a boolean vector in the data frame with a specific
prefix column encoding) contains an element (stored under a specific
column name). For example, it can be used to check if a candidate set
contains the failed component.
}
