% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_set_size}
\alias{md_set_size}
\title{md_set_size}
\usage{
md_set_size(md, setvar = "x", cname = NULL)
}
\arguments{
\item{md}{A masked data frame.}

\item{setvar}{A character string representing the column prefix for matrix encoding. Default is "x".}

\item{cname}{An optional character string for the name of the new column denoting the size of the sets. Defaults to concatenating "size_", and `setvar`.}
}
\value{
A decorated masked data frame with a new column `cname` indicating the size of the sets.
}
\description{
Decorates a masked data frame by adding a new column indicating the size of
sets (represented as boolean vectors with a specific prefix column encoding).
}
