% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_unmark_latent}
\alias{md_unmark_latent}
\title{Mark a variable in a masked data frame as latent.}
\usage{
md_unmark_latent(md, vars)
}
\arguments{
\item{md}{masked data to modify}

\item{vars}{variables to mark as latent.}
}
\description{
Mark a variable in a masked data frame as latent.
}
