% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_write_as_json}
\alias{md_write_as_json}
\title{helper function}
\usage{
md_write_as_json(file, ls)
}
\arguments{
\item{file}{connection}

\item{ls}{object (typically list) to write as json

TODO: recursively search x and only write those values that can be
      serialized and deserialized by jsonlite.}
}
\description{
helper function
}
