% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_write_csv_with_meta}
\alias{md_write_csv_with_meta}
\title{Write data frame object to a CSV (comma separated file), optionally
with associated attribute data (stored as JSON in comments)}
\usage{
md_write_csv_with_meta(df, file, comment = "#")
}
\arguments{
\item{df}{a data frame with attributes to write}

\item{file}{filename for csv}

\item{comment}{denotes a comment block}
}
\description{
Write data frame object to a CSV (comma separated file), optionally
with associated attribute data (stored as JSON in comments)
}
