% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md}
\alias{md}
\title{Constructor for masked data.}
\usage{
md(x)
}
\arguments{
\item{x}{object to convert to masked data.}
}
\description{
Takes an object \code{x} and converts it to masked data, a tibble (data
frame) with some extra attributes, e.g., \code{latent} attribute to specify
which variables are latent in the model.
}
