% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_latent}
\alias{md_latent}
\title{Obtain a list of latent variables from masked data.}
\usage{
md_latent(md)
}
\arguments{
\item{md}{masked data to retrieve latent variables from.}
}
\description{
Obtain a list of latent variables from masked data.
}
