% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_unmark_latent}
\alias{md_unmark_latent}
\title{md_unmark_latent}
\usage{
md_unmark_latent(md, vars)
}
\arguments{
\item{md}{masked data to modify}

\item{vars}{character vector containing variables to unmark as latent.}
}
\value{
modified masked data with updated latent variables
}
\description{
Unmark variables in a masked data frame as latent.
}
