% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{md_write_csv_with_meta}
\alias{md_write_csv_with_meta}
\title{md_write_csv_with_meta}
\usage{
md_write_csv_with_meta(df, file, comment = "#", ...)
}
\arguments{
\item{df}{a data frame with attributes to write (like a masked data frame)}

\item{file}{filename for csv}

\item{comment}{denotes a comment block}

\item{...}{additional arguments to pass}
}
\description{
Write data frame object to a CSV (comma separated file), optionally
with associated attribute data (stored as JSON in comments)
}
