% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{print.tbl_md}
\alias{print.tbl_md}
\title{Print method for masked data (\code{tbl_md}).}
\usage{
\method{print}{tbl_md}(x, drop_latent = F, ...)
}
\arguments{
\item{x}{masked data to print}

\item{drop_latent}{Boolean, drop the latent random variables}

\item{...}{additional arguments to pass}
}
\description{
Print method for masked data (\code{tbl_md}).
}
